Updates for the state machine logic including burnwire and networking; pressure updates; fix IPC pipes; small updates.
-- Updating state machine logic.
---- Update the burnwire logic to avoid triggering based on stale data; it now saves a timeout start time to a file upon the first dive of a deployment.
---- No longer using the ST_DEPLOY state; directly go to recording at surface or depth based on pressure readings.
---- Update network logic; previously, it would prevent the state from advancing to recording until the WiFi grace period had passed.
---- Move logic for forcing networking off so it is not also tied to the pressure sensor being enabled.
---- Check for low battery voltage while at the surface in addition to depth.
---- Stop data capture once the battery becomes critical.

-- Record raw pressure readings instead of using an initial reading as calibration.

-- Adjust the default dive and surface pressure thresholds to avoid triggering the state change based on noisy readings or surface activity.

-- Fix the IPC response pipe path, so commands can be successfully processed
---- Was previously trying to use the command pipe as the response pipe so was infinitely hanging when processing a command.

-- Abort recovery functions early if the board is uninitialized.

-- Rename some variables for clarity.
-- Updates to address minor build warnings.

+++ packages/ceti-tag-data-capture/src/sensors/ecg_helpers/ecg_adc.h
@@ -117,7 +117,7 @@ void ecg_adc_config_reset();
 void ecg_adc_config_apply();
 void ecg_adc_write_config_register(uint8_t data);
 uint8_t ecg_adc_read_register(uint8_t reg);
-int ecg_adc_raw_read_data(void);
+long ecg_adc_raw_read_data(void);
 int ecg_adc_read_data(int* exit_flag, long long timeout_us);
 int ecg_adc_read_data_ready();
